1.) Changed type of numPartsTo(Send|Recv) to int as MPI_Get_count uses no unsigend ints. 2.) Changed declaration of countMolecules interface as we can have a lot of molecules. (See also declaration of the molecule ID)
git-svn-id: https://scm.projects.hlrs.de/svn/ls1/MarDyn/trunk@181 a63bd714-7e14-4b5e-94e7-302c8c8ff188

+++ src/md_io/PartGen.cpp
@@ -389,7 +389,7 @@ unsigned long PartGen::readPhaseSpace(
   createClusters(particleContainer, bBoxMin, bBoxMax, domain, domainDecomp);
 
   vector<Component>& dcomponents = domain->getComponents();
-  vector<int> partsPerComp;
+  vector<unsigned long> partsPerComp;
   partsPerComp.resize(_numberOfComponents);
   domain->setglobalNumMolecules(domainDecomp->countMolecules(particleContainer, partsPerComp));
   for(unsigned int i=0; i<partsPerComp.size(); i++) {
